stx-distcloud: API ref doc content added.
I added the api-ref-dcmanager-v1.rst file, which is content
for the API reference manual to the api-ref/source dir.
This represents the converted old-style files to the newer
OpenStack supported RST files.

I updated the index.rst file to include the new .rst file
so that the api-ref document can build.

Change-Id: I76f6ee2658efced7038e0e5b659902382fd72284
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>