DC fix audit sync failure
During the sync Security Group needs to be compared with the tenant
names and not tenant ids. The tenant id which is the uuid are different
between regions. The compare does not match and an attempt to create
the Security Group fails since it already exists. This commit changes
the compare to use tenant names.

Closes-Bug: 1802397

Change-Id: I9363945868b857802e137dcf05757ec3691230c1
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>